Fixed mmult. Added functor combinators
Matrix multiplication was wrong. The type was no general enough.
We could not multiply together a 2x3 and 3x5 matrix and get a 2x5.
This has been fixed now. (Also the order of matrix arguments was
wrong in old definition of mmult.)

Also added functor combinators. They are not used yet.